test: split admin promotion tests
Changed:
Split the admin promotions page test suite into focused files with shared test setup.

Benefit:
Future promotion page changes are easier to review and verify without editing one oversized test file.
